GS-845: Moved `process()` logic to `bulk_session_manager_parent.php`.
