+++ classes/bulk_session_manager_sitelevel.php
@@ -12,7 +12,6 @@
 
 namespace mod_facetoface;
 
-use DateTime;
 use stdClass;
 
 defined('MOODLE_INTERNAL') || die();
@@ -70,207 +69,42 @@ class bulk_session_manager_sitelevel extends bulk_session_manager_parent {
         $this->validate_common_fields($record, $index);
     }
 
-    /**
-     * Process valid records to create sessions.
-     * Assumes validate() has already been called to check correctness.
-     *
-     * @return bool true on success, false if any errors occurred
-     */
-    public function process(): bool {
-        global $DB;
+    protected function process_record(array $record, int $index, array $customfieldsbyshortname): void {
+        $shortname = trim($record['Course Shortname']);
+        $f2fname = trim($record['Face-to-Face Activity Name']);
+        $matched = $this->match_records($shortname, $f2fname);
+        $course = $matched['course'];
+        $f2frecord = $matched['facetoface'];
 
-        $allcustomfields = facetoface_get_session_customfields();
-        $customfieldsbyshortname = [];
+        if (!$course) {
+            $this->errors[] = [
+                $index,
+                get_string('error:coursenotfound', 'facetoface', $shortname)];
 
-        foreach ($allcustomfields as $field) {
-            $customfieldsbyshortname[strtolower($field->shortname)] = $field;
+            return;
         }
 
-        foreach ($this->records as $index => $record) {
-            $session = new stdClass();
+        if (!$f2frecord) {
+            $this->errors[] = [
+                $index,
+                get_string(
+                    'error:f2fnotfound',
+                    'facetoface',
+                    (object)[
+                        'shortname' => $shortname,
+                        'f2fname'   => $f2fname
+                    ]
+                )
+            ];
 
-            $shortname = trim($record['Course Shortname']);
-            $f2fname = trim($record['Face-to-Face Activity Name']);
-            $matched = $this->match_records($shortname, $f2fname);
-            $course = $matched['course'];
-            $f2frecord = $matched['facetoface'];
-
-            if (!$course) {
-                $this->errors[] = [
-                    $index,
-                    get_string('error:coursenotfound', 'facetoface', $shortname)];
-
-                continue;
-            }
-
-            if (!$f2frecord) {
-                $this->errors[] = [
-                    $index,
-                    get_string(
-                        'error:f2fnotfound',
-                        'facetoface',
-                        (object)['shortname' => $shortname,
-                            'f2fname'   => $f2fname]
-                    )
-                ];
-
-                continue;
-            }
-
-            $session->facetoface = $f2frecord->id;
-
-            $session->datetimeknown = 1;
-            if (
-                isset($record['Session Date/Time Known']) &&
-                $record['Session Date/Time Known'] === 'no'
-            ) {
-                $session->datetimeknown = 0;
-            }
-
-            $session->starttime = null;
-            if (
-                !empty($record['Start Date']) &&
-                !empty($record['Start Time'])
-            ) {
-                $session->starttime = strtotime(str_replace('/', '-', $record['Start Date'].' '.$record['Start Time']));
-            }
-
-            $session->finishtime = null;
-            if (
-                !empty($record['Finish Date']) &&
-                !empty($record['Finish Time'])
-            ) {
-                $session->finishtime = strtotime(str_replace('/', '-', $record['Finish Date'].' '.$record['Finish Time']));
-            }
-
-            if (
-                $session->datetimeknown &&
-                (empty($session->starttime) ||
-                    empty($session->finishtime))
-            ) {
-                $this->errors[] = [
-                    $index,
-                    get_string('error:invaliddatetimedata', 'facetoface')
-                ];
-
-                continue;
-            }
-
-            $session->allowcancellations = 1;
-            if (
-                isset($record['Allow Cancellations']) &&
-                $record['Allow Cancellations'] === 'no'
-            ) {
-                $session->allowcancellations = 0;
-            }
-
-            $session->capacity = 10;
-            if (
-                isset($record['Capacity']) &&
-                is_numeric($record['Capacity'])
-            ) {
-                $session->capacity = (int)$record['Capacity'];
-            }
-
-            $session->allowoverbook = 1;
-            if (
-                isset($record['Allow Overbookings']) &&
-                $record['Allow Overbookings'] === 'no'
-            ) {
-                $session->allowoverbook = 0;
-            }
-
-            $session->duration = 0;
-            if (
-                isset($record['Duration']) &&
-                is_numeric($record['Duration'])
-            ) {
-                $session->duration = (int)$record['Duration'];
-            }
-
-            $session->normalcost = 0;
-            if (
-                isset($record['Normal Cost']) &&
-                is_numeric($record['Normal Cost'])
-            ) {
-                $session->normalcost = $record['Normal Cost'];
-            }
-
-            $session->discountcost = 0;
-            if (
-                isset($record['Discount Cost']) &&
-                is_numeric($record['Discount Cost'])
-            ) {
-                $session->discountcost = $record['Discount Cost'];
-            }
-
-            $session->details = '';
-            if (
-                isset($record['Details']) &&
-                is_string($record['Details'])
-            ) {
-                $session->details = $record['Details'];
-            }
-
-            $session->timecreated = time();
-            $session->timemodified = time();
-
-            $sessionid = $DB->insert_record('facetoface_sessions', $session);
-            if (!$sessionid) {
-                $this->errors[] = [
-                    $index,
-                    get_string('error:failedtocreatesession', 'facetoface')
-                ];
-
-                continue;
-            }
-
-            // Insert session dates.
-            $sessionsdate = new stdClass();
-            $sessionsdate->sessionid = $sessionid;
-            $sessionsdate->timestart = $session->starttime;
-            $sessionsdate->timefinish = $session->finishtime;
-            $sessionsdateid = $DB->insert_record('facetoface_sessions_dates', $sessionsdate);
-
-            if (!$sessionsdateid) {
-                $this->errors[] = [
-                    $index,
-                    get_string('error:failedtocreatedates', 'facetoface', $sessionid)
-                ];
-            }
-
-            foreach ($record as $column => $value) {
-                // If the column does not start with "Customfield_", skip it.
-                if (strpos($column, 'Customfield_') !== 0) {
-
-                    continue;
-
-                }
-
-                $shortname = strtolower(substr($column, strlen('Customfield_')));
-
-                // If we don’t have a matching custom field for $shortname, skip it.
-                if (!isset($customfieldsbyshortname[$shortname])) {
-                    $this->errors[] = [
-                        $index,
-                        get_string('error:unknowncustomfieldshort', 'facetoface', $shortname)
-                    ];
-
-                    continue;
-                }
-
-                // Otherwise, save the custom field.
-                $field = $customfieldsbyshortname[$shortname];
-                if (!facetoface_save_customfield_value($field->id, $value, $sessionid, 'session')) {
-                    $this->errors[] = [
-                        $index,
-                        get_string('error:couldnotsavecustomfieldshort', 'facetoface', $shortname)
-                    ];
-                }
-            }
+            return;
         }
 
-        return empty($this->errors);
+        $session = new stdClass();
+        $session->facetoface = $f2frecord->id;
+
+        // Use common session creation logic.
+        $this->process_session_record($session, $record, $index, $customfieldsbyshortname);
     }
 
     /**
